pciutils: Add hard dependency on wget
Even though nothing in core is supposed to be downloading anything from
the Internet as part of its build (because if you're building an ISO, by
definition you'll have an incomplete system), pciutils thinks that it's
special and insists on downloading a new pci.ids file just in case it's
changed in the last half hour or so.

(The pci.ids file chagnes on average once a month. It's not that
special.)

So until there's a better fix, just add another layer of stupid, ugly
hackwork on top of the existing pile of problems.